Don't discuss hexadecimal when introducing `xxd`.
The point of this section of the write-up is to describe a file's
contents. Most people are constantly told "everything in a computer is
just a bunch of zeros and ones." But that's not what you'll see if you
open the canonical representation of file contents in a hex editor. The
leap from "a series of zeros and ones" to "hexadecimal representation of
bytes" is the missing link I'm trying to explain. To do that, one must
*not* introduce hexadecimal until one first understands what the ones
and zeroes actually are.