Precisify Numerical Nomenclature
`0` and `1` are not the first two integers:  integers are the numbers $latex \ldots -1,-2,0,1,2,\ldots$.  They go on infinitely in both directions, so there is no "first integer" or "second integer".  Rather, `0` and `1` are the first two digits.  Also `0` and `1` in this context are not numerals.  A numeral is a sequence of symbols that represents a number, like `10`, which represents the number two.  Here `0` and `1` are being used not as numerals, but as the building blocks that compose numerals.  The `1` in `10` does not represent a number as long as it appears in `10`.

To be really precise, the phrase "base-2 number" should also be replaced by the phrase "base-2 numeral".  But that might be misleading, I'm not sure.

See [this](https://revolutionarymathematics.wordpress.com/2017/01/15/numbers-and-numerals/) for more information about the distinctions between numbers, numerals, and digits.